requirements.txt
python-telegram-bot==11.1.0
requests==2.21.0

